// src/store/action-types/anime-action-types.ts
export enum AnimeActionTypes{
    SET_ANIMES = "SET_ANIMES",
    SET_ANIME_ERROR = "SET_ANIME_ERROR",
    SET_RANDOM_ANIMES = "SET_RANDOM_ANIME",
    SET_PAGINATED_ANIMES = "SET_PAGINATED_ANIMES",
    SET_LAST_ANIME_PAGE = "SET_LAST_ANIME_PAGE",
    SET_IS_ANIME = "SET_IS_ANIME",
    SET_ANIME_SEARCH_RESULT = "SET_ANIME_SEARCH_RESULT",
    SET_ANIME_SEARCH_VALUE = "SET_ANIME_SEARCH_VALUE",
    SET_ANIME_GENRES = "SET_ANIME_GENRES",
    SET_ANIME_BY_GENRE = "SET_ANIME_BY_GENRE",
    SET_CURRENT_ANIME_TITLE = "SET_CURRENT_ANIME_TITLE",
    SET_CURRENT_ANIME_PAGE = "SET_CURRENT_ANIME_PAGE",
    SET_POPULAR_ANIME = "SET_POPULAR_ANIME",
    SET_ANIME_REVIEW = "SET_ANIME_REVIEW "
}